Move Tracking class into it's own file
